Extract MetricsProperty
