--- java/core/src/main/java/sleeper/core/properties/instance/DashboardProperty.java
@@ -1,53 +1,0 @@
-/*
- * Copyright 2022-2024 Crown Copyright
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package sleeper.core.properties.instance;
-
-import sleeper.core.properties.SleeperPropertyIndex;
-import sleeper.core.properties.validation.SleeperPropertyValueUtils;
-
-import java.util.List;
-
-/**
- * Definitions of instance properties relating to the metrics dashboard.
- */
-public interface DashboardProperty {
-    UserDefinedInstanceProperty DASHBOARD_TIME_WINDOW_MINUTES = Index.propertyBuilder("sleeper.dashboard.time.window.minutes")
-            .description("The period in minutes used in the dashboard.")
-            .defaultValue("5")
-            .validationPredicate(SleeperPropertyValueUtils::isPositiveInteger)
-            .propertyGroup(InstancePropertyGroup.DASHBOARD)
-            .runCdkDeployWhenChanged(true).build();
-
-    static List<UserDefinedInstanceProperty> getAll() {
-        return Index.INSTANCE.getAll();
-    }
-
-    /**
-     * An index of property definitions in this file.
-     */
-    class Index {
-        private Index() {
-        }
-
-        private static final SleeperPropertyIndex<UserDefinedInstanceProperty> INSTANCE = new SleeperPropertyIndex<>();
-
-        private static UserDefinedInstancePropertyImpl.Builder propertyBuilder(String propertyName) {
-            return UserDefinedInstancePropertyImpl.named(propertyName)
-                    .addToIndex(INSTANCE::add);
-        }
-    }
-}
