+++ java/core/src/main/java/sleeper/core/properties/instance/UserDefinedInstanceProperty.java
@@ -62,7 +62,7 @@ public interface UserDefinedInstanceProperty extends InstanceProperty {
             index.addAll(GarbageCollectionProperty.getAll());
             index.addAll(CompactionProperty.getAll());
             index.addAll(QueryProperty.getAll());
-            index.addAll(DashboardProperty.getAll());
+            index.addAll(MetricsProperty.getAll());
             index.addAll(LoggingLevelsProperty.getAll());
             index.addAll(AthenaProperty.getAll());
             index.addAll(DefaultProperty.getAll());
