+++ scripts/templates/instanceproperties.template
@@ -131,26 +131,6 @@ sleeper.task.runner.memory.mb=1024
 # This must be >0 and <= 900.
 sleeper.task.runner.timeout.seconds=900
 
-# The namespaces for the metrics used in the metrics stack.
-sleeper.metrics.namespace=Sleeper
-
-# The reserved concurrency for the table metrics lambda.
-# See reserved concurrency overview at:
-# https://docs.aws.amazon.com/lambda/latest/dg/configuration-concurrency.html
-# sleeper.metrics.concurrency.reserved=
-
-# The maximum given concurrency allowed for the table metrics lambda.
-# See maximum concurrency overview at:
-# https://aws.amazon.com/blogs/compute/introducing-maximum-concurrency-of-aws-lambda-functions-when-using-amazon-sqs-as-an-event-source/
-sleeper.metrics.concurrency.max=10
-
-# The number of tables to calculate metrics for in a single invocation. This will be the batch size
-# for a lambda as an SQS FIFO event source. This can be a maximum of 10.
-sleeper.metrics.batch.size=1
-
-# Whether to calculate table metrics for offline tables.
-sleeper.metrics.offline.enabled=false
-
 # If true, properties will be reloaded every time a long running job is started or a lambda is run.
 # This will mainly be used in test scenarios to ensure properties are up to date.
 sleeper.properties.force.reload=false
@@ -1211,7 +1191,27 @@ sleeper.default.query.results.page.size=131072
 sleeper.query.warm.lambda.period.minutes=5
 
 
-## The following properties relate to the dashboard.
+## The following properties relate to metrics.
+
+# The namespaces for the metrics used in the metrics stack.
+sleeper.metrics.namespace=Sleeper
+
+# The reserved concurrency for the table metrics lambda.
+# See reserved concurrency overview at:
+# https://docs.aws.amazon.com/lambda/latest/dg/configuration-concurrency.html
+# sleeper.metrics.concurrency.reserved=
+
+# The maximum given concurrency allowed for the table metrics lambda.
+# See maximum concurrency overview at:
+# https://aws.amazon.com/blogs/compute/introducing-maximum-concurrency-of-aws-lambda-functions-when-using-amazon-sqs-as-an-event-source/
+sleeper.metrics.concurrency.max=10
+
+# The number of tables to calculate metrics for in a single invocation. This will be the batch size
+# for a lambda as an SQS FIFO event source. This can be a maximum of 10.
+sleeper.metrics.batch.size=1
+
+# Whether to calculate table metrics for offline tables.
+sleeper.metrics.offline.enabled=false
 
 # The period in minutes used in the dashboard.
 sleeper.dashboard.time.window.minutes=5
